Move `LocalLiveHackrNewsLoader` and `LiveHackrNewsStore` to production
